Identify app storage keys based off store identity
`UserDefaults.init(suiteName:)` returns a different object each time,
even for the same suite name, and so our current hashing for app storage
keys can fail for folks that re-fetch the suite, _e.g._ in a computed
property.

This PR attempts to identify the underlying suite to get around this.
Unfortunately it does involve performing a selector that is private, and
may run afoul of App Store review, but it's worth considering this
option.

An alternative path from here would be to bracket use in `#if DEBUG` and
instead detect when a the same suite is passed with a different object
and runtime warn.