Add `$shared.load(key)`
While `try await $shared = Shared(key)` is a convenient way to suspend
when reconfiguring a key to fetch a new value, because it only fully
replaces the value on a successful load, it can lead to shared
properties being left in an undesirable state, and it can lead to lost
errors, since the `loadError` is not populated on the original
`$shared`.

To avoid these gotchas, let's introduce an API that is responsible for
the assignment and loading of the key, so that assignment is always
successful.